external/opal-prd: Use libflash/blocklevel for MTD/file accesses
pnor.c existed before blocklevel, it is time that this code got updated to
use the available libraries.

Changes include using the blocklevel accessors for MTD reads and writes
rather than read() and write() on a file descriptor.

This patch also makes use of the arch_flash_init() auto detection of the
/dev/mtd device which corresponds to the MTD device exposed by skiboot for
access to platform flash.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>