hw/bt: allow BT driver to use different buffer size
The buffer size used by the BT driver defaults to 64bytes even when
the BMC reports a different size. This patch removes the restriction
as there is no apparent reason for it.

The extra byte added on the input and output buffer size in the test
also seems superfluous. The buffer lengths used later in the code are
correct as they take into account the extra byte needed for the
message size :

	bt.caps.input_buf_len = msg->data[1] + 1;
	bt.caps.output_buf_len = msg->data[2] + 1;

>From IPMI Specs:

    22.10 Get BT Interface Capabilities Command

    For Bytes 3 and 4 (Input and Output Buffer size), the buffer
    message size is the largest value allowed in first byte (length
    field) of any BT request or response message. For a send, this
    means if Get BT Interface Capabilities returns 255 in byte 3
    (input buffer size) the driver can actually write 256 bytes to the
    input buffer (adding one for the length byte (byte 1) that is sent
    in with the request.)

Signed-off-by: Cédric Le Goater <clg@fr.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>