external/gard: Fix displaying 'cleared' gard records
When a garded component is replaced hostboot detects this and updates the
gard partition.

What hostboot does is set the record_id field to 0xFFFFFFFF but leaves the
rest of the flash untouched, this has caused issues with the gard tool the
thinking was that an entire record of all 0xFF bytes would signal not a
valid record.

This patch add rectifies this issue and `gard list` will no longer show any
record with an id of 0xFFFFFFFF.

Equivalent to 91cebdcd557dfc2dfc8050a34edb11a43facaf74 in master

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>