nvlink: Add fence mode emulation for NPUs
As NPUs are emulated PCI devices, they do not get physically fenced as real
PCI devices do.  As such, when the device is in a state that it should be
fenced, we need to emulate this behaviour by returning all 1s in config
space reads.

This will be utilised by error injection in subsequent patches.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Acked-By: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>