nvlink: Add freeze and fence error injection
Enable NPU freeze and fence injection through debugfs.

For example, if a NPU is PCI bus 8, a freeze on PE 1 can be injected with:

echo 1:0:0:0:0 >> /sys/kernel/debug/powerpc/PCI0008/err_injct

or a fence on PE 2 on PCI bus 9 with:

echo 2:1:0:0:0 >> /sys/kernel/debug/powerpc/PCI0009/err_injct

These will cause the appropriate EEH event to occur upon a DMA to the
NVLink.

PE number was added to the npu_dev struct to enable this.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Acked-By: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>