hw/phb3: Ensure PQ bits are cleared in the IVC when masking IRQ
When we mask an interrupt, we may race with another interrupt coming
in from the hardware.  If this occurs, the P and/or Q bit may end up
being set but we never EOI/clear them.  This could result in a lost
interrupt or the next interrupt that comes in after re-enabling never
being presented.

This patch ensures that when masking an interrupt, any pending P/Q
bits are cleared.

This fixes a bug seen with some CAPI workloads which have lots of
interrupt masking at the same time as high interrupt load.  The fix is
not specific to CAPI though.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Tested-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>