Enable Simics UART console
The UART is a simulated ns16550 with memory mapped registers.

A /simics dt node is detected and a SIMICS_QUIRK is added to chip quirks
similar to MAMBO_CALLOUTS.  It can contain an ns16550 dt node with a property
console-bar.

The LPC UART code is reused and this will work without an LPC bus in the model.

Signed-off-by: Ryan Grimm <grimm@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>