ipmi-sel: Fix esel event logger to handle early boot PANIC events
We pre-allocate IPMI message for PANIC event and use that memory to send
PANIC event to BMC. Presently we return NULL if we have not initiated PANIC
event message. So we won't be able to log early failure events.

This patch tries to initialize ipmi message instead of returning NULL.

Also intialize elog before ipmi_sel_init. Otherwise we will not be able
to create elog message.

Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>