xscom: Return OPAL_WRONG_STATE on XSCOM ops if CPU is asleep
xscom_read and xscom_write return OPAL_SUCCESS if they worked, and
OPAL_HARDWARE if they didn't.  This doesn't provide information about why
the operation failed, such as if the CPU happens to be asleep.

This is specifically useful in error scanning, so if every CPU is being
scanned for errors, sleeping CPUs likely aren't the cause of failures.

So, return OPAL_WRONG_STATE in xscom_read and xscom_write if the CPU is
sleeping.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Reviewed-by: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>