hmi: Only raise a catchall HMI if no other components have
Regardless of whether a handler for a specific component has raised an
event to deal with a HMI or not, skiboot will raise an extra HMI at the
end of the detection.  This is problematic, as if one component reports
it is recoverable but another reports it is not, the last handler to be
called will have priority.

Rework this to instead only send a HMI event if no handler has raised an
event themselves.  This will send an unknown, unrecoverable HMI event
since the cause could not be found.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Reviewed-by: Alistair Popple <alistair@popple.id.au>
Reviewed-by: Mahesh Salgaonkar <mahesh@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>