hmi: Add handling for NPU checkstops
If the NPU detects an unrecoverable error, it will send a HMI.  This is
problematic since unhandled HMIs will checkstop the entire system, which
is not the intended behaviour of a NPU failure.  Instead, the NPU
emulated PCI devices should be fenced as part of EEH.

Add support for handling NPU HMIs.  This works by finding the NPU
responsible for the HMI, checking its error registers, and sending a
recoverable HMI event.  The NPU itself cannot actually recover, but the
system should not be brought down.  Fence mode is set on the NPU, such
that any further operations on the NPU will trigger EEH, and it will be
subsequently fenced from the system.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Reviewed-by: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>