hdata: Define SPIRA-H and SPIRA-S structure
In latest firwmare the SPIRA is split into two portions, a hypervisor
initialized and managed portion and a service processor (FSP) built portion,
to accommodate secure boot. The hypervisor portion will be referred to as
the SPIRA-H and the FSP portion referred to as the SPIRA-S.

The legacy N-tuples that are no longer being used have been removed. N-tuples
that are initialized by the hypervisor at compile/build time but not modified
by the FSP reside in the SPIRA-H. All other N-tuples reside in the SPIRA-S that
is built dynamically by the FSP at the beginning of that HDAT memory pointed to
by the “Host Data Area” N-tuple in SPIRA-H.

This patch defines new SPIRA-H and SPIRA-S structures.

No functionality change.

Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>