pflash: Allow building under yocto.
The makefiles under external/* utilize the $(CROSS_COMPILE) variable
to determine the cross-compiler prefix.  In a few places,
$(CROSS_COMPILE)gcc is called instead of $(CC).  The issue with this is
that yocto build passes some compile flags as part of $(CC) instead of
$(CFLAGS), the most important of these is '--sysroot=...'.  Without the
proper --sysroot flag, pflash compile fails to find critical libc
headers like stdio.h.

This change delegates setting of $(CC) and $(LD) to
external/common/rules.mk, which is widely used in the external tree, and
ensures that:
    1) $(CC) is used instead of $(CROSS_COMPILE)gcc.
    2) CC is only set when not passed from the environment.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Reviewed-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>