Make console-log time more readable: seconds rather than timebase
I've tried to align it with what the kernel prints today

The existing logs show:

[20287269125,5] PSI[0x000]: Found PSI bridge [working=1, active=0]
[890425679,5] BT: Interface initialized, IO 0x00e4
[1249199642,4] SLW: HB-provided idle states property found
[1249344409,5] NVRAM: Size is 576 KB
[1484422964,3] NVRAM: Layout appears sane

The mftb() output is not very meaningful, the changed output shows:

[   38.315784591,5] CENTAUR:   FSI host: 0x0 cMFSI0 port 2
[   38.315922971,5] PSI[0x000]: Found PSI bridge [working=1, active=0]
[    1.448765255,5] BT: Interface initialized, IO 0x00e4
[    2.398136129,5] NVRAM: Size is 576 KB
[    3.145017865,3] NVRAM: Layout appears sane

For the output sample, I've taken bits where one can see the time move back
w.r.t. previous log, but that was always the case. I don't think that is
worth fixing here

Adds an additional divide and modulo for every log printed. I've also
fixed the test cases that run as a part of make check

Signed-off-by: Balbir Singh <bsingharora@gmail.com>
[stewart@linux.vnet.ibm.com: fix up tb in console tests]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>