phb3: Refactor to move capi enable to its own function
The phb3_set_capi_mode function was getting a little long, and the
enable procedure will need to be parameterised to support CAPP DMA mode,
so split it out into it's own function in preparation for this. Also
replace the multiple if (mode == xxx) cases with an equivalent switch
construct since it will shortly gain yet another case.

No functional changes.

Signed-off-by: Ian Munsie <imunsie@au1.ibm.com>
Reviewed-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
[stewart@linux.vnet.ibm.com: enabled The Oatmeal apostrophe compatibility mode]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>