phb: Work around XSL bug sending PTE updates with wrong scope
The XSL incorrectly sends out PTE updates with nodal scope instead of
group scope, which causes a system checkstop. This is exacerbated by the
fact that recent Linux kernels do not set the R bit in the hashed page
tables, requiring the XSL to write them back to set that bit. Work
around this issue by forcing all commands to be unlimited scope.

This might have a slight performance impact, but it is expected to be
negligible since the XSL is only used for translations, not for block
data transfers. To avoid impacting other cards, tie it off the use of
DMA mode, as only the XSL based Mellanox CX4 card uses this mode and is
affected.

Signed-off-by: Ian Munsie <imunsie@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>