nvlink: Present chip ID as the NPU PHB slot location
EEH errors in the kernel report the physical slot location of the
erroneous PE and its PHB.  For NPU devices, the PE's slot location will
refer to the physical GPU the link is associated with, and the PHB is
actually a NPU chip which has no relevance to a physical slot on a board.

Rather than reporting N/A for a NPU PHB's slot location, present the chip
number.  It's not particularly useful, but better than nothing.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>