nvlink: Enable NPU device BAR before triggering freeze
NPU freeze injection works by performing an invalid MMIO read on the NPU
device BAR.  If the BAR isn't enabled, which is the case when the
appropriate driver isn't loaded, this checkstops the machine.

Work around this by making sure the BAR is enabled before performing the
read.  The idea of an error inject doing anything other than an error
inject isn't great, but it's better than unintentionally crashing your
machine.

Also, fix the comment incorrectly stating the operation was a write
instead of a read.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>