nvlink: Print error message when NPU is fenced
NPU fences aren't recoverable, and as such, would require user
intervention to have a working system again.  The fence will be picked up
by the kernel through EEH, but this doesn't happen until the NPU is used
for something.  So, let's print a message so it's obvious when this
happens.

A helper function was added to reduce duplication.  This also enables code
in skiboot to un-fence a NPU, which is useful to NPU developers but very
stupid otherwise.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>