fwts/generate-fwts-olog: Fix regex matches to allow more printf formats
We currently aren't handling the 'll' length modifier, or widths
starting with zero (eg %016x). The former is becase we have the shorter
'l' match in the group first (so borking on the second l). The latter is
because we don't allow a leading zero on width matches.

This change fixes those issues.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>