PHB3: Disable UTL outbound discard error interrupt
The UTL outbound discard error (bit#12) in UTL_PCIE_PORT_IRQ_EN
(offset: 0x558) isn't set in initial setup. It's set wrongly after
a reset on root port. With this bit set, frozen (all) error was
observed on the PHB to which a LPFC adapter is connected directly.

This removes the bit in reset handler to avoid the unexpected
frozen (all) error.

 BZ: 142877

Reported-by: Pridhiviraj Paidipeddi <ppaidipe@in.ibm.com>
Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>