Produce XZ compressed skiboot.lid as part of build
With a recent HostBoot change, we can have an XZ compressed PAYLOAD
that's automagically detected (looking at magic numbers).

This gives us three great benefits:
1) it's transparent, uncompressed skiboot.lid works everywhere
2) it lets us grow greater than 1MB binary, as long as we compress
   down to <1MB
3) It speeds up boot. We currently compress down to 230kb rather than
   922kb, which is much quicker to read off flash.

This patch produces skiboot.lid.xz alongside standard skiboot.lid.
We currently use crc32 as this is supported by hostboot. Future HB
may support crc64, but this seems to be disabled currently.

Having CRC32 in the PAYLOAD partition also gives the advantage of
error detection in PAYLOAD, which we previously did not have as it
was not an ECC protected partition.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>