flash: Rework error paths and messages for multiple flash controllers
The current flash code was written with only one flash chip, which is
a system_flash (ie. the PNOR image), in mind.

Now that we have mambo bogusdisk flash, we can have many flash chips.
This is resulting in some confusing output messages.

This reworks some of the error paths and warnings to make this more
coherent when we have multiple flash chips.

We assume everything can be a system flash, so I've removed the
is_system_flash parameter from flash_register().  We'll use the first
system flash we find and warn if we find another since discovery order
is not a guaranteed API.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Acked-by: Jeremy Kerr <jk@ozlabs.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>