Stop adding legacy linux, phandle to device tree, just add phandle
Since the very first commit of a skeleton powernv platform in Linux,
the kernel has been able to deal with phandle rather than linux,phandle.

Thus, every kernel in the real world that has a hope of even getting to
a twinkle in the eye of booting knows about phandle.

So, we don't need to add linux,phandle along with phandle as it's just
redundantly redundant.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>