test/hello_world: always use shutdown type zero
The hello world kernel fails to correctly set r3 before making the
shutdown opal call. On FSP machines only shutdown types 0 and 1 are
recognised as valid shutdown types. If any other type is specified
(in r3) the call is rejected with an OPAL_PARAMETER error and the
machine will continue running.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>