hmi: Clean up NPU FIR debug messages
With the skiboot log set to debug, the FIR (and related registers) were
logged all in the same message.  It was too much for one line, didn't
clarify if the numbers were in hex, and didn't show leading zeroes.

So, split it into two lines, with leading zeroes and a "0x" prefix.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>